requirements.txt
# Core Data Science Libraries
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2

# Advanced Machine Learning Models
xgboost==2.0.3
lightgbm==4.1.0
imbalanced-learn==0.11.0

# Deep Learning (Optional - for Neural Networks)
tensorflow==2.15.0
keras==2.15.0

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# Web Framework
Flask==3.0.0
Werkzeug==3.0.1
Flask-CORS==4.0.0

# Environment Variables
python-dotenv==1.0.0

# Google Generative AI (for Chatbot)
google-generativeai==0.3.2

# Model Persistence
joblib==1.3.2

# Logging and Monitoring
python-json-logger==2.0.7

# API Documentation
flask-swagger-ui==4.11.1

# Data Processing
scipy==1.11.4

# Jupyter Notebook Support (optional)
jupyter==1.0.0
ipykernel==6.27.1

# Progress bars
tqdm==4.66.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0

